refactor(utils): restructure package
