+++ packages/KTable/KTable.vue
@@ -96,7 +96,7 @@
 <script>
 import KEmptyState from '@kongponents/kemptystate/KEmptyState.vue'
 import KSkeleton from '@kongponents/kskeleton/KSkeleton.vue'
-import { useRequest, useDebounce } from '../utils/lib/utils'
+import { useRequest, useDebounce } from '../utils/utils'
 
 import Vue from 'vue'
 import VueCompositionAPI, { computed, defineComponent, onMounted, ref, watch } from '@vue/composition-api'
