--- packages/utils/__tests__/utils.test.js
@@ -1,7 +1,0 @@
-'use strict'
-
-// const utils = require('..')
-
-describe('utils', () => {
-  it('needs tests', () => {})
-})
